requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.8.0

# Twitter API
tweepy>=4.14.0

# OpenAI (optional - use Ollama as free alternative)
openai>=1.0.0

# PDF generation
weasyprint>=60.0
reportlab>=4.0.0

# Web scraping and parsing
lxml>=4.9.0
html5lib>=1.1
pytrends>=4.9.0

# Data validation
pydantic>=2.5.0

# YAML configuration
pyyaml>=6.0

# Database
# SQLite is built into Python, no extra package needed

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Optional: Local LLM support
# ollama (install separately: pip install ollama)
